# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple chat backend (messages history in Redis).",
        "title": "Chat API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
        "/messages": {
            "get": {
                "description": "Return last N messages (chronological order)",
                "produces": [
                    "application/json"
                ],
                "summary": "List recent messages",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Items count (1..HISTORY_LIMIT)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Message"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new chat message and store it in Redis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create message",
                "parameters": [
                    {
                        "description": "Message payload",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.createMessage"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.Message": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "isSystem": {
                    "type": "boolean"
                },
                "sendtime": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "httpserver.createMessage": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}